pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gcp-auditor"
version = "0.4.1"
description = "Fast, opinionated AWS security scanner. Curated checks. Zero noise. Copy-paste fixes."
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "GCP Auditor Team" },
]
keywords = ["aws", "cloud", "audit", "security", "compliance", "cis", "scanner", "devops", "terraform"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
]
dependencies = [
    "google-api-python-client>=2.118.0",
    "google-auth>=2.28.0",
    "google-auth-httplib2>=0.2.0",
    "typer>=0.15.0",
    "rich>=13.9.0",
    "pydantic>=2.10.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=6.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[project.scripts]
gcp-auditor = "cloud_audit.cli:app"

[project.urls]
Homepage = "https://github.com/abdullahkamil/gcp-auditor"
Documentation = "https://github.com/abdullahkamil/gcp-auditor#readme"
Repository = "https://github.com/abdullahkamil/gcp-auditor"
Issues = "https://github.com/abdullahkamil/gcp-auditor/issues"
Changelog = "https://github.com/abdullahkamil/gcp-auditor/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/cloud_audit"]

[tool.ruff]
target-version = "py39"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "SIM", "TCH"]
ignore = ["S101"]

[tool.ruff.lint.per-file-ignores]
"src/cloud_audit/providers/*/checks/*.py" = ["E501", "S112"]
"src/cloud_audit/cli.py" = ["TC003"]
"src/cloud_audit/providers/gcp/provider.py" = ["UP037"]
"tests/**" = ["TC003"]

[tool.mypy]
python_version = "3.10"
strict = true

[[tool.mypy.overrides]]
module = [
    "googleapiclient.*",
    "google.cloud.*",
    "google.auth.*",
    "google.oauth2.*",
    "google.auth.credentials",
    "yaml",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
